Only scroll to next page on first filter
